+++ quirrel/src/client/buildtime-checks.ts
